requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-allauth==0.43.0
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
oauthlib==3.2.0
pipenv==2022.4.8
platformdirs==2.5.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
